fix(domain/service/mapper): converte TipoUsuario para entidade JPA + corrige repositories e services
- TipoUsuario: enum -> @Entity JPA com id, descricao, ativo, timestamps (tabela tipo_usuario ja existia no banco)
- TipoUsuarioMapper: volta a ser interface MapStruct gerenciando TipoUsuario como entidade
- TipoUsuarioService: restaura CRUD completo usando TipoUsuarioRepository
- TipoUsuarioRepository: JpaRepository<TipoUsuario, Integer> (ja existia correto)
- AnexoRepository: Long -> Integer (Anexo.id e Integer)
- AuditoriaRepository: Integer -> UUID (Auditoria.id e UUID gerado por GenerationType.UUID)
- AuditoriaService: Integer -> UUID em todos os metodos
- ClienteTipoMapper: tipoUsuario.getId() (agora valido pois TipoUsuario voltou a ser entidade)
- EquipamentoService: remove 2 null extras na chamada searchIncludingInativos (repositorio aceita 4 params)
- InsumoFornecedorRepository: JpaRepository<InsumoFornecedor, InsumoFornecedorAux> -> JpaRepository<InsumoFornecedor, InsumoFornecedorAuxId>
- LogAlteracaoAuxRepository: findByTipoAlteracao(String) -> findByTipoAlteracao(TipoAlteracao) (enum)
- ImagemProjetoMapper: substitui expression java() invalida por @Named helper usando Projeto.ofId()

+++ pruma/src/main/java/com/br/pruma/core/repository/AnexoRepository.java
@@ -3,5 +3,5 @@ package com.br.pruma.core.repository;
 import com.br.pruma.core.domain.Anexo;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AnexoRepository  extends JpaRepository<Anexo, Long> {
+public interface AnexoRepository extends JpaRepository<Anexo, Integer> {
 }
